# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This are the Smylet APP APIs.",
        "title": "Smylet API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/references/universities": {
            "get": {
                "description": "List all universities",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Universities"
                ],
                "summary": "List universities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "University name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "University code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "University city",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "University country",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "University state",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reference.UniversitySerializer"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "reference.UniversitySerializer": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}